ZonedDateTime & Instant methods moved to reminderItems
